Move myers diff util
