--- vscode/src/core-diff/diff/test-examples/fastapi-tabs-vs-spaces.py.diff
@@ -1,42 +1,0 @@
-from fastapi import FastAPI
-
-app = FastAPI()
-
-@app.get("/")
-def read_root():
-\treturn {"Hello": "World"}
-
-@app.get("/landing")
-def read_landing_page():
-\treturn {"message": "Welcome to the landing page"}
-
----
-
-from fastapi import FastAPI, HTTPException
-
-app = FastAPI()
-
-@app.get("/")
-def read_root():
-    return {"Hello": "World"}
-
-@app.get("/landing")
-def read_landing_page():
-    raise HTTPException(status_code=404, detail="Page not found")
-
----
-
-- from fastapi import FastAPI
-+ from fastapi import FastAPI, HTTPException
- 
- app = FastAPI()
- 
- @app.get("/")
- def read_root():
-- \treturn {"Hello": "World"}
-+     return {"Hello": "World"}
- 
- @app.get("/landing")
- def read_landing_page():
-+     raise HTTPException(status_code=404, detail="Page not found")
-- \treturn {"message": "Welcome to the landing page"}
